refactor: Extract write/ss to its own file
To remove dependency on Pair of _writer.js
